Added list on AddPordukt Page, Changed eanNumber to Long, Added funct to get produkt by EanNumber. Added funct for Handleliste.

+++ app/src/main/java/com/example/stockpalapp/model/PantryProduct.kt
@@ -7,7 +7,7 @@ data class PantryProduct(
     @DocumentId val id: String = "",
     override val name: String = "",
     override val number: Int = 0,
-    val eanNumber: Double = 0.0,
+    val eanNumber: Long = 0,
     val category: String = "",
     val image: String = "",
     val expDate: Timestamp? = null,
